requirements/dev.txt
-r base.txt
django-extensions
flower

